More upgrading utils / testing. Sharded some effects and big utils files.
