+++ src/engine/core/managers/upgrades/index.ts
@@ -1,2 +1,3 @@
 export * from "@/engine/core/managers/upgrades/UpgradesManager";
 export * from "@/engine/core/managers/upgrades/utils";
+export * from "@/engine/core/managers/upgrades/item_upgrades_types";
