+++ src/engine/core/managers/upgrades/utils/index.ts
@@ -1,3 +1,4 @@
 export * from "@/engine/core/managers/upgrades/utils/upgrades_get";
 export * from "@/engine/core/managers/upgrades/utils/upgrades_install";
+export * from "@/engine/core/managers/upgrades/utils/upgrades_label_utils";
 export * from "@/engine/core/managers/upgrades/utils/upgrades_price_utils";
